Refactor & cleanup effects
